feat: add support for both header and footer.
Transclusion has been removed and directive `stepContainer` is now required in order to provide load form steps. This also allow full control of markup structure.

BREAKING CHANGE: prior to this, there could only be a header or a footer. Now there can be both, as long as an element with the `stepContainer` directive is inside the `multiStepContainer` directive.

Fixes #2